# Conflicts:
#	src/main/java/com/wanted/backend/global/aspect/LoggingAspect.java

+++ src/main/java/com/wanted/backend/domain/community/application/command/CreateReviewCommand.java
@@ -3,7 +3,7 @@ package com.wanted.backend.domain.community.application.command;
 public record CreateReviewCommand(
         Long memberId,
         Long courseId,
-        Double rating,
+        Integer rating,
         String content
 ) {
 }
